third_party/WebKit/LayoutTests/http/tests/devtools/sources/outline-javascript-1-expected.txt

Verify javascript outline


Running: testSimpleFunction
Text:
    function foo(a, b, c) {}
Outline:
    foo(a, b, c):0:9

Running: testSpreadOperator
Text:
    function foo(a, b, ...c) {}
Outline:
    foo(a, b, ...c):0:9

Running: testVariableDeclaration
Text:
    var a = function(a,b) { }
Outline:
    a(a, b):0:4

Running: testMultipleVariableDeclaration
Text:
    var a = function(a,b) { }, b = function(c,d) { }
Outline:
    a(a, b):0:4
    b(c, d):0:27

Running: testObjectProperty
Text:
    a.b.c = function(d, e) { }
Outline:
    c(d, e):0:0
